Split code into page and form
The page handles submission, the form handles layout and controller components
